Refactoring. Moved all fixtures to conftest.py for their convenient reusability
